+++ src/Modal/Modal.uni.driver.js
@@ -1,7 +1,10 @@
 import { baseUniDriverFactory } from 'wix-ui-test-utils/base-driver';
 
 export const modalUniDriverFactory = (base, body) => {
-  const getPortal = () => body.$('.portal');
+  const getPortal = async () => {
+    const dataHook = await base.attr('data-hook');
+    return dataHook ? body.$(`.portal.portal-${dataHook}`) : body.$('.portal');
+  };
   const getOverlay = () => body.$('.ReactModal__Overlay');
   const getContent = () => body.$('.ReactModal__Content');
   const getCloseButton = () => body.$('[data-hook="modal-close-button"]');
@@ -10,22 +13,21 @@ export const modalUniDriverFactory = (base, body) => {
 
   return {
     ...baseUniDriverFactory(base),
+
+    /** true if the modal is on the DOM */
+    exists: async () => !!(await getPortal()),
+
     /** true when the module is open */
     isOpen,
     /** true if theme <arg> exists in the modal */
-    isThemeExist: theme =>
-      getPortal()
-        .$(`.${theme}`)
-        .exists(),
-    getChildBySelector: async selector =>
-      (await getPortal()
-        .$(selector)
-        .exists())
-        ? getPortal().$(selector)
-        : null,
+    isThemeExist: async theme => (await getPortal()).$(`.${theme}`).exists(),
+    getChildBySelector: async selector => {
+      const portal = await getPortal();
+      return (await portal.$(selector).exists()) ? portal.$(selector) : null;
+    },
     /** true if the modal is scrollable */
     isScrollable: async () =>
-      !(await getPortal().hasClass('portalNonScrollable')),
+      !(await (await getPortal()).hasClass('portalNonScrollable')),
     closeButtonExists: () => getCloseButton().exists(),
     /** click on the modal overlay (helpful for testing if the modal is dismissed) */
     clickOnOverlay: () => getOverlay().click(),