+++ src/Tag/Tag.js
@@ -87,6 +87,7 @@ class Tag extends React.PureComponent {
         [styles.withThumb]: thumb,
         [styles.disabled]: disabled,
         [styles.clickable]: onClick !== noop,
+        [styles.hoverable]: !disabled,
       },
     );
   }