+++ src/Timeline/test/Timeline.visual.js
@@ -69,6 +69,19 @@ const tests = [
           ],
         },
       },
+      {
+        it: 'no suffix', // prop variation (e.g. small)
+        props: {
+          items: [
+            {
+              label: 'simple text',
+            },
+            {
+              label: 'some long text '.repeat(24),
+            },
+          ],
+        },
+      },
     ],
   },
 ];